chore: moved revoke page to inside identity
