+++ src/pages/errors/index.js
@@ -1,9 +1,7 @@
 import Page500 from './500';
 import Page404 from './404';
-import Revoked from './revoked';
 
 export {
     Page500,
     Page404,
-    Revoked,
 };
