+++ src/pages/identity/revoked/Revoked.css
@@ -23,9 +23,7 @@
     }
 
     & .avatar {
-        margin: 0 auto;
         margin-bottom: 2.3rem;
-        display: block;
         fill: var(--color-tan);
     }
